Moved AsyncSingleArticleRecommendationProvider back to article_recommendation
